upgrade nextjs template
