--- templates/rescript-template-nextjs/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "next/babel"
-  ],
-  "plugins": [
-  ]
-}
