--- templates/rescript-template-nextjs/bsconfig.json
@@ -1,24 +1,0 @@
-{
-  "name": "rescript-template-nextjs",
-  "namespace": false,
-  "reason": { "react-jsx": 3 },
-  "refmt": 3,
-  "bs-dependencies": ["@rescript/react"],
-  "ppx-flags": [],
-  "sources": [
-    { "dir": "src", "subdirs": true }
-  ],
-  "package-specs": {
-    "module": "es6",
-    "in-source": true
-  },
-  "suffix": ".mjs",
-  "warnings": {
-    "number": "-3",
-    "error": "+101+8"
-  },
-  "gentypeconfig": {
-    "language": "untyped",
-    "shims": []
-  }
-}
